Check for prestige getting behind to lower bonus map levels.
Maps will run for world bonus if damage or health is low. The level run is the max of oneshot level, min siphon level, minimum drop level, or 6.  Minimum drop level was removed if there were no drops available. Now, it's also removed if all prestiges of the needed stat (atk/hp) are available for upgrade. This prevents end of run slowness taking on max level maps if not necessary.