Improved heliumHistory/heliumLog.
Trimps used to not save the time of last portal, so Trimpz just used the time from when the script started or the last portal if it was running.  Now that Trimps has the data, Trimpz uses it to make the "totalHeliumPerHour" accurate regardless of when the script is started. So, you'll be able to see an accurate value for the run you start the script on.