Redid all damage and health calculations. Fixes Titimp dmg issue.
Calculate all damage and health for soldiers and enemies instead of getting some of it from the HTML.  This allows Titimp damage to not mislead the script.  This also paves the wsy for doing Repeat Maps since that will require calculating these things while running maps (which is/was never done).