Stabilize geneticists more.
To help prevent the crazy flip-flopping from 0 to full.  This is caused by buying housing when the total breed time for the next set of soldiers is high (happens when soldiers are alive a long time), which makes it look like it's taking a very long time to breed.  It only fires now concerning total breed time if the time remaining is more than 1 and soldier health is less than half.