Added Last Cell setting to alleviate Javascript issues.
From the tooltip:
Begin Toxic or Void after this cell in the zone. Later is better for Toxic and Void, but too late causes an issue when Javascript does its inevitable pausing, Trimps will play instant catch up based on time lapse on Javascript resuming, and Trimpz will not have a chance to see whatever cells were skipped.  This can lead to void or toxic being completely skipped if the cell is too late combined with a long enough Javascript pause.